ager-fincas-app/requirements.txt
django==4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.1
psycopg2==2.9.2
pytz==2021.3
gunicorn==20.1.0
django-heroku==0.3.1